Move build script to Makefile, update circle
The build started failing recently. Not sure why. We don't have any
actual tests on this repo, we just build the image and make sure that
doesn't fail. We can do that via a makefile, to be more consistent with
our other engine repos.

Made a few small tweaks to the Dockerfile to get the build compatible
with Circle's version of Docker. Seems like we can't copy a file into a
dir without including a trailing slash or the full path of the target
file, or it turns the dir into a file.

Discussed a bit here:
https://discuss.circleci.com/t/failed-docker-build-at-a-copy-step-in-dockerfile/5440/4