Bump Go version to 1.12.5
Current build configuration uses Go 1.9.3, which is neither "fresh"
not supported anymore. This commit updates the used version to 1.12.5
(together with bumping Alpine version to 3.9), which should apply
support for more recent syntax, linting rules etc.

Also the import for `lint` package was change to where it is currently
available.

Additionally this commit adds few small changes in the Dockerfile, to
align it with the one that is present in codeclimate-gofmt project.